refactor(cfg): move block data types to separate file (#6319)
Pure refactor. No logic has been changed.
